+++ server/v2/appmanager/store/store.go
@@ -4,7 +4,6 @@ import (
 	"sync/atomic"
 
 	"cosmossdk.io/core/store"
-	corestore "cosmossdk.io/core/store"
 )
 
 var _ Store = (*Storage[Database])(nil)
@@ -65,10 +64,10 @@ func (s state[DB]) Has(key []byte) (bool, error) { return s.db.Has(s.storeKey, s
 
 func (s state[DB]) Get(bytes []byte) ([]byte, error) { return s.db.Get(s.storeKey, s.version, bytes) }
 
-func (s state[DB]) Iterator(start, end []byte) (corestore.Iterator, error) {
+func (s state[DB]) Iterator(start, end []byte) (store.Iterator, error) {
 	return s.db.Iterator(s.storeKey, s.version, start, end)
 }
 
-func (s state[DB]) ReverseIterator(start, end []byte) (corestore.Iterator, error) {
+func (s state[DB]) ReverseIterator(start, end []byte) (store.Iterator, error) {
 	return s.db.ReverseIterator(s.storeKey, s.version, start, end)
 }