+++ tests/integration/bank/keeper/deterministic_test.go
@@ -86,6 +86,7 @@ func initDeterministicFixture(t *testing.T) *deterministicFixture {
 		addresscodec.NewBech32Codec(sdk.Bech32MainPrefix),
 		sdk.Bech32MainPrefix,
 		authority.String(),
+		nil,
 	)
 
 	blockedAddresses := map[string]bool{
@@ -150,24 +151,34 @@ func TestGRPCQueryBalance(t *testing.T) {
 		coin := getCoin(rt)
 		fundAccount(f, addr, coin)
 
-		req := banktypes.NewQueryBalanceRequest(addr, coin.GetDenom())
+		addrStr, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addr)
+		assert.NilError(t, err)
+
+		req := banktypes.NewQueryBalanceRequest(addrStr, coin.GetDenom())
 
 		testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.Balance, 0, true)
 	})
 
+	addr1Str, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addr1)
+	assert.NilError(t, err)
+
 	fundAccount(f, addr1, coin1)
-	req := banktypes.NewQueryBalanceRequest(addr1, coin1.GetDenom())
+	req := banktypes.NewQueryBalanceRequest(addr1Str, coin1.GetDenom())
 	testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.Balance, 1087, false)
 }
 
 func TestGRPCQueryAllBalances(t *testing.T) {
 	t.Parallel()
 	f := initDeterministicFixture(t)
+	addressCodec := codectestutil.CodecOptions{}.GetAddressCodec()
 
 	rapid.Check(t, func(rt *rapid.T) {
 		addr := testdata.AddressGenerator(rt).Draw(rt, "address")
 		numCoins := rapid.IntRange(1, 10).Draw(rt, "num-count")
 		coins := make(sdk.Coins, 0, numCoins)
+
+		addrStr, err := addressCodec.BytesToString(addr)
+		assert.NilError(t, err)
 
 		for i := 0; i < numCoins; i++ {
 			coin := getCoin(rt)
@@ -178,7 +189,7 @@ func TestGRPCQueryAllBalances(t *testing.T) {
 
 		fundAccount(f, addr, coins...)
 
-		req := banktypes.NewQueryAllBalancesRequest(addr, testdata.PaginationGenerator(rt, uint64(numCoins)).Draw(rt, "pagination"), false)
+		req := banktypes.NewQueryAllBalancesRequest(addrStr, testdata.PaginationGenerator(rt, uint64(numCoins)).Draw(rt, "pagination"), false)
 		testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.AllBalances, 0, true)
 	})
 
@@ -188,7 +199,10 @@ func TestGRPCQueryAllBalances(t *testing.T) {
 	)
 
 	fundAccount(f, addr1, coins...)
-	req := banktypes.NewQueryAllBalancesRequest(addr1, nil, false)
+	addr1Str, err := addressCodec.BytesToString(addr1)
+	assert.NilError(t, err)
+
+	req := banktypes.NewQueryAllBalancesRequest(addr1Str, nil, false)
 
 	testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.AllBalances, 357, false)
 }
@@ -199,6 +213,8 @@ func TestGRPCQuerySpendableBalances(t *testing.T) {
 
 	rapid.Check(t, func(rt *rapid.T) {
 		addr := testdata.AddressGenerator(rt).Draw(rt, "address")
+		addrStr, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addr)
+		assert.NilError(t, err)
 
 		// Denoms must be unique, otherwise sdk.NewCoins will panic.
 		denoms := rapid.SliceOfNDistinct(rapid.StringMatching(denomRegex), 1, 10, rapid.ID[string]).Draw(rt, "denoms")
@@ -213,10 +229,10 @@ func TestGRPCQuerySpendableBalances(t *testing.T) {
 			coins = sdk.NewCoins(append(coins, coin)...)
 		}
 
-		err := banktestutil.FundAccount(f.ctx, f.bankKeeper, addr, coins)
+		err = banktestutil.FundAccount(f.ctx, f.bankKeeper, addr, coins)
 		assert.NilError(t, err)
 
-		req := banktypes.NewQuerySpendableBalancesRequest(addr, testdata.PaginationGenerator(rt, uint64(len(denoms))).Draw(rt, "pagination"))
+		req := banktypes.NewQuerySpendableBalancesRequest(addrStr, testdata.PaginationGenerator(rt, uint64(len(denoms))).Draw(rt, "pagination"))
 		testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.SpendableBalances, 0, true)
 	})
 
@@ -228,7 +244,10 @@ func TestGRPCQuerySpendableBalances(t *testing.T) {
 	err := banktestutil.FundAccount(f.ctx, f.bankKeeper, addr1, coins)
 	assert.NilError(t, err)
 
-	req := banktypes.NewQuerySpendableBalancesRequest(addr1, nil)
+	addr1Str, err := codectestutil.CodecOptions{}.GetAddressCodec().BytesToString(addr1)
+	assert.NilError(t, err)
+
+	req := banktypes.NewQuerySpendableBalancesRequest(addr1Str, nil)
 	testdata.DeterministicIterations(t, f.ctx, req, f.queryClient.SpendableBalances, 1777, false)
 }