+++ x/auth/ante/ante.go
@@ -42,8 +42,8 @@ func NewAnteHandler(options HandlerOptions) (sdk.AnteHandler, error) {
 	anteDecorators := []sdk.AnteDecorator{
 		NewSetUpContextDecorator(options.AccountKeeper), // outermost AnteDecorator. SetUpContext must be called first
 		NewExtensionOptionsDecorator(options.ExtensionOptionChecker),
-		NewValidateBasicDecorator(options.AccountKeeper),
-		NewTxTimeoutHeightDecorator(options.AccountKeeper),
+		NewValidateBasicDecorator(options.AccountKeeper.Environment()),
+		NewTxTimeoutHeightDecorator(options.AccountKeeper.Environment()),
 		NewValidateMemoDecorator(options.AccountKeeper),
 		NewConsumeGasForTxSizeDecorator(options.AccountKeeper),
 		NewDeductFeeDecorator(options.AccountKeeper, options.BankKeeper, options.FeegrantKeeper, options.TxFeeChecker),