+++ x/distribution/client/common/common_test.go
@@ -7,10 +7,15 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec/legacy"
+	codectestutil "github.com/cosmos/cosmos-sdk/codec/testutil"
 )
 
 func TestQueryDelegationRewardsAddrValidation(t *testing.T) {
-	clientCtx := client.Context{}.WithLegacyAmino(legacy.Cdc)
+	cdcOpts := codectestutil.CodecOptions{}
+	clientCtx := client.Context{}.
+		WithLegacyAmino(legacy.Cdc).
+		WithAddressCodec(cdcOpts.GetAddressCodec()).
+		WithValidatorAddressCodec(cdcOpts.GetValidatorCodec())
 
 	type args struct {
 		delAddr string